+++ projects/demo-maps/src/app/components/example-group-legend/example-group-legend.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ExampleGroupLegendComponent } from './example-group-legend.component';
+
+describe('ExampleGroupLegendComponent', () => {
+  let component: ExampleGroupLegendComponent;
+  let fixture: ComponentFixture<ExampleGroupLegendComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ ExampleGroupLegendComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(ExampleGroupLegendComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});