+++ projects/demo-maps/src/app/components/example-layer-description/example-layer-description.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ExampleLayerDescriptionComponent } from './example-layer-description.component';
+
+describe('ExampleLayerDescriptionComponent', () => {
+  let component: ExampleLayerDescriptionComponent;
+  let fixture: ComponentFixture<ExampleLayerDescriptionComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ ExampleLayerDescriptionComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(ExampleLayerDescriptionComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});